+++ src/Cocoar.Configuration.HttpPolling/HttpPollingProviderQueryOptions.cs
@@ -7,19 +7,15 @@ public sealed class HttpPollingProviderQueryOptions : IProviderQuery
     public string UrlPathOrAbsolute { get; }
     // Selecting a property (section) from the fetched JSON payload, optional (colon-separated path)
     public string? ConfigurationPath { get; }
-    // Optionally wrap the resulting object into this path
-    public string? TargetPath { get; }
     public IReadOnlyDictionary<string, string>? Headers { get; }
 
     public HttpPollingProviderQueryOptions(
-        string urlPathOrAbsolute,
-        string? configurationPath = null,
-        string? targetPath = null,
-        IReadOnlyDictionary<string,string>? headers = null)
+    string urlPathOrAbsolute,
+    string? configurationPath = null,
+    IReadOnlyDictionary<string,string>? headers = null)
     {
         UrlPathOrAbsolute = urlPathOrAbsolute ?? throw new ArgumentNullException(nameof(urlPathOrAbsolute));
         ConfigurationPath = configurationPath;
-        TargetPath = targetPath;
         Headers = headers;
     }
 }