+++ src/Cocoar.Configuration/Providers/FileSourceProvider/FileSourceProviderQueryOptions.cs
@@ -5,6 +5,5 @@ namespace Cocoar.Configuration.Providers.FileSourceProvider;
 public record FileSourceProviderQueryOptions(
 	string Filename,
 	string? ConfigurationPath = null,
-	string? TargetPath = null,
 	TimeSpan? DebounceTime = null
 ) : IProviderQuery;